requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.1
discord.py==2.3.2
ffmpeg-python==0.2.0
frozenlist==1.4.1
future==1.0.0
gTTS==2.5.1
idna==3.6
multidict==6.0.5
pycparser==2.21
pydub==0.25.1
PyNaCl==1.5.0
pypiwin32==223
python-dotenv==1.0.1
pyttsx3==2.90
pywin32==306
requests==2.31.0
urllib3==2.2.1
yarl==1.9.4
